# Conflicts:
#	README.md
#	src/main/java/pl/pabilo8/immersiveintelligence/api/crafting/ElectrolyzerRecipe.java
#	src/main/java/pl/pabilo8/immersiveintelligence/client/gui/overlay/gun/GuiOverlayGunBase.java
#	src/main/java/pl/pabilo8/immersiveintelligence/client/util/amt/AMTQuads.java
#	src/main/java/pl/pabilo8/immersiveintelligence/common/IICreativeTab.java
#	src/main/java/pl/pabilo8/immersiveintelligence/common/compat/TerrafirmaHelper.java
#	src/main/java/pl/pabilo8/immersiveintelligence/common/crafting/IIRecipes.java
#	src/main/java/pl/pabilo8/immersiveintelligence/common/item/tools/ItemIITrenchShovel.java
#	src/main/java/pl/pabilo8/immersiveintelligence/common/weaponsystem/IIWeaponUpgrade.java
#	src/main/java/pl/pabilo8/immersiveintelligence/common/weaponsystem/weapons/IIWeaponMachinegun.java

+++ src/main/java/pl/pabilo8/immersiveintelligence/common/item/armor/ItemIILightEngineerChestplate.java
@@ -174,7 +174,7 @@ public class ItemIILightEngineerChestplate extends ItemIILightEngineerArmorBase 
 			return;
 
 		ElectricDamageSource dmg = (ElectricDamageSource)dSource;
-		dmg.dmg = (hasUpgrade(s, "anti_static_mesh")&&p.isInWater())?dmg.dmg*LightEngineerArmor.antiStaticMeshWaterDamageMod: 0;
+		dmg.dmg = (hasUpgrade(s, "anti_static_mesh")&&p.isInWater())?(float)(dmg.dmg*LightEngineerArmor.antiStaticMeshWaterDamageMod): 0;
 	}
 
 	@Override
